Add aarch64 assembly optimization (ARMv8a 64 bits)
The fixmuldiv functions don't need inline assembly to be fast
in this architecture; the compiler (both clang and GCC) figure
out to use the optimal instructions for this (which is 2 instruction
sequence), and when letting the compiler emit the instructions
instead of using inline assembly, the compiler is able to
interleave those instructions with other instructions,
improving scheduling, making it even faster than when using
inline assembly.

Overall, this gives about 50% speedup.

With this in place, one could probably remove the -Wno-#warnings
flag from Android.mk as well.

Change-Id: Idd585ff57891464dcb1e173974702d0cbbd967c4
Signed-off-by: mydongistiny <jaysonedson@gmail.com>